+++ SOD.Common/ModFolderContent/CHANGELOG.md
@@ -5,6 +5,7 @@
 - Added method "UpdateConfigFileLayout" to PluginController for IPluginBindings clean up support
 - Added Dialog Helper to create dialogs between player and npcs. (Lib.Dialog)
 - Fixed DDS entries (names, descriptions, text) not being loaded for custom sync disks when loading/creating a new/existing game.
+- Added PlayerStatus helper, it provides helpers to set status on player such as illegal status.
 
 **1.1.7**
 - Made sync disk re-raise events optional (extra parameter in builder, default true)