+++ SOD.Common/ModFolderContent/CHANGELOG.md
@@ -6,6 +6,7 @@
 - Added Dialog Helper to create dialogs between player and npcs. (Lib.Dialog)
 - Fixed DDS entries (names, descriptions, text) not being loaded for custom sync disks when loading/creating a new/existing game.
 - Added PlayerStatus helper, it provides helpers to set status on player such as illegal status.
+- Sync disk DDS records are now unique and cannot be overriden anymore by other mods by accident
 
 **1.1.7**
 - Made sync disk re-raise events optional (extra parameter in builder, default true)