+++ SOD.Common/ModFolderContent/CHANGELOG.md
@@ -7,6 +7,7 @@
 - Added PlayerStatus helper, it provides helpers to set status on player such as illegal status. (Lib.PlayerStatus)
 - Fixed DDS entries (names, descriptions, text) not being loaded for custom sync disks when loading/creating a game from an already ongoing game.
 - Sync disk DDS records are now unique and cannot be overriden anymore by other mods by accident
+- Added ChainLoader util to detect the presence of another mod (Lib.PluginDetection)
 
 **1.1.7**
 - Made sync disk re-raise events optional (extra parameter in builder, default true)