+++ SOD.Common/ModFolderContent/CHANGELOG.md
@@ -1,4 +1,8 @@
 # CHANGELOG
+**1.1.9**
+- Added extra operators (<, >, <=, >=) to TimeData struct in Lib.Time
+- Fixed parsing issue with GetFnvHashCode extension
+
 **1.1.8**
 - Added non-generic bind extension on BepInEx ConfigFile object
 - Added support for enum values in interface bindings